requirements.txt
pytest~=6.0.2
flask~=1.1.2
Flask-SQLAlchemy~=2.4.4